+++ MA/Relationship.cpp
@@ -8,6 +8,6 @@ Relationship::Relationship(std::string n) : Name(n) {
 Relationship::~Relationship() {
 }
 
-void Relationship::SetScoringFunction(std::function<int (LineWrap *, LineWrap *)> aScoringFunction) {
+void Relationship::SetScoringFunction(std::function<double (LineWrap *, LineWrap *)> aScoringFunction) {
     ScoringFunction = aScoringFunction;
 }
